requirements.txt
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
celery==5.1.0
redis==3.5.3
